Реализовал логику обновления карточки фильма
