+++ src/utils/common.js
@@ -1,5 +1,3 @@
-import Abstract from '../view/abstract.js';
-
 export const getRandomInteger = (a = 0, b = 1) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
@@ -46,20 +44,16 @@ export const getRandomElementOfArray = (list) => {
 
 export const numberInRange = (count, rangeMin, rangeMax) => count >= rangeMin && count <= rangeMax;
 
-export const replace = (newChild, oldChild) => {
-  if (oldChild instanceof Abstract) {
-    oldChild = oldChild.getElement();
+export const updateItem = (items, update) => {
+  const index = items.findIndex((item) => item.id === update.id);
+
+  if (index === -1) {
+    return items;
   }
 
-  if (newChild instanceof Abstract) {
-    newChild = newChild.getElement();
-  }
-
-  const parent = oldChild.parentElement;
-
-  if (parent === null || oldChild === null || newChild === null) {
-    throw new Error('Can\'t replace unexisting elements');
-  }
-
-  parent.replaceChild(newChild, oldChild);
+  return [
+    ...items.slice(0, index),
+    update,
+    ...items.slice(index + 1),
+  ];
 };
